// front/src/pages/My_Profile/Profile_Page/Other_Profile/interface.ts
export interface data_of_player {
    username: string;
    level: number;
    total_games: number;
    win_games: number;
    lose_games: number;
    points: number;
    profile_image: string;
    cover_image: string;
    win_1_game:boolean;
    win_3_games:boolean;
    win_10_games:boolean;
    win_30_games:boolean;
    reach_level_5:boolean;
    reach_level_15:boolean;
    reach_level_30:boolean;
    perfect_win_game:boolean;
    perfect_win_tournaments:boolean;
  }